💘 提高代码质量

+++ src/interaction.cpp
@@ -8,12 +8,12 @@ Interaction::Interaction(const long interval_ms)
 
 void Interaction::update_interaction_tick()
 {
-	last_interaction_tick = millis();
+	last_interaction_tick = static_cast<long>(millis());
 }
 
 bool Interaction::can_interact() const
 {
-	return ((long)(millis()) - last_interaction_tick) >= interact_interval_ms;
+	return (static_cast<long>(millis()) - last_interaction_tick) >= interact_interval_ms;
 }
 
 void Interaction::add_cooldown(const long cooldown_ms)
